Refactor helpers and airtable files in utils.
